Substitute git archive for uae of tar directly
In my code I used tar to make a deployable copy of the 7App directory which contains the code to prepare a tax return. git uses the archive command to do the same thing. I decided to switch to the git archive command to see if it had any advantages. The git archive command is enclosed in the makeArchive.sh file.